+++ src/cpp/qos/DDSQosPolicies.cpp
@@ -124,12 +124,12 @@ bool PartitionQosPolicy::addToCDRMessage(CDRMessage_t* msg)
 	for(std::vector<std::string>::iterator it = names.begin();it!=names.end();++it)
 	{
 		this->length +=4;
-		this->length += it->size()+1;
-		rest = (it->size()+1)%4;
+		this->length += (uint16_t)it->size()+1;
+		rest = ((uint16_t)it->size()+1)%4;
 		this->length += rest!=0?rest:4-rest;
 	}
 	valid &= CDRMessage::addUInt16(msg, this->length);//this->length);
-	valid &= CDRMessage::addUInt32(msg,this->names.size());
+	valid &= CDRMessage::addUInt32(msg,(uint32_t)this->names.size());
 	for(std::vector<std::string>::iterator it = names.begin();it!=names.end();++it)
 		valid &= CDRMessage::addString(msg,*it);
 	//valid &= CDRMessage::addOctetVector(msg,&name);
